Provide an initial version of authot fly-out metadata
On the search results page, when hovering author names, display a box with a
link to extra information about that author. The information is then taken
from the new PersonalAuthorProperty model and displayed according their rank
(defined in the PersonalAuthorPropertyRank model).

Added a custom/hacked CSS style to display the hover, this will be later
refactored/improved with the new UI revamp.

Use default alt for images.