Provide a command to process and store author metadata and properties
This branch provides a new Django management command to process and store
document personal author's metadata into an intermediary new model. The
metadata and properties are taken from the already defined helper that would
manipulate and massage the information in the unmanaged PersonalAuthorProperty
"frozen" model.

The command would also download and store in the configured Djanfo store the
main image for every author, if defined.