take 3 (#147)
pr processing looks good

When a PR is opened, a PR-specific solr image will be used for testing.  This does not involve pushing new version data to the repository--the targets in ./justfile work with the `prNum` environment variable (defined in workflow) to include the PR number in the image tag (e.g.: v0.0.0-[PR#]-solr)

  * if no new data files submitted to the PR branch
     * if pr-specific solr image exists
        * pull & retag as solr
     * if parent-branch solr image exists
        * pull & retag as `solr`
     * if parent-branch solr image does not exist
        * build it and push it
        * retag as `solr`
        * yes, we're in a PR, but if the image doesn't exist, it can't have been used for a deployment
  * if a new data file is committed to the PR
        * build it and push it

The retagging of the image (`solr`) allows for setting the tag to what is expected in `docker-compose.ci.yaml`.  Whether its a parent branch's image or a pr-specific image, it gets renamed to `$( prNum= just tag )-solr`

`$( prNum= just tag )` produces `nuremberg:vM.m.p-solr`