Document details: conditionally show citation links (#179)
The transcript citation links are now conditionally shown depending on the
information from the `tblCasesDatesList`. When the value of the column
`TranscriptCitationSeqNo` is NULL, citation links are shown as before.
When 0, citation links are suppressed, and when not NULL and non zero,
a link to the transcript with reference to the sequence number is provided.