Reduce replicas to 1
There is an odd minification/compression bug related to static assets.
Each nuremberg container seems to generate a different hashed filename
which breaks CSS/JS loading if there is more than one replica in play.

Need to resolve this soon in case of getting hit with a ton of requests
and for smooth rollouts.